requirements.txt
torch
sentence-transformers
pymilvus
pdfminer.six
numpy
python-dotenv
tqdm